+++ src/main/java/io/github/hyshmily/hotkey/actuator/HotKeyEndpoint.java
@@ -1,14 +1,29 @@
+/*
+ * Copyright 2026 Hyshmily. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.github.hyshmily.hotkey.actuator;
 
 import com.github.benmanes.caffeine.cache.Cache;
-import io.github.hyshmily.hotkey.hotkeycache.HotKeyProperties;
 import io.github.hyshmily.hotkey.algorithm.Item;
 import io.github.hyshmily.hotkey.algorithm.TopK;
+import io.github.hyshmily.hotkey.hotkeycache.HotKeyProperties;
+import io.github.hyshmily.hotkey.hotkeycache.SingleFlight;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 import org.springframework.boot.actuate.endpoint.annotation.Endpoint;
 import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 
@@ -17,17 +32,17 @@ public class HotKeyEndpoint {
 
   private final TopK hotKeyDetector;
   private final Cache<String, Object> caffeineCache;
-  private final Cache<String, CompletableFuture<Object>> inflightLoads;
+  private final SingleFlight singleFlight;
   private final int l1MaxSize;
 
   public HotKeyEndpoint(
       TopK hotKeyDetector,
       Cache<String, Object> caffeineCache,
-      Cache<String, CompletableFuture<Object>> inflightLoads,
+      SingleFlight singleFlight,
       HotKeyProperties properties) {
     this.hotKeyDetector = hotKeyDetector;
     this.caffeineCache = caffeineCache;
-    this.inflightLoads = inflightLoads;
+    this.singleFlight = singleFlight;
     this.l1MaxSize = properties.getLocalCacheMaxSize();
   }
 
@@ -48,7 +63,7 @@ public class HotKeyEndpoint {
     info.put("l1CacheSize", caffeineCache.estimatedSize());
     info.put("l1MaxSize", l1MaxSize);
 
-    info.put("inflightSize", inflightLoads.estimatedSize());
+    info.put("inflightSize", singleFlight.estimatedInflightSize());
 
     info.put("recentlyExpelled",
       hotKeyDetector.expelled().stream()