+++ src/main/java/io/github/hyshmily/hotkey/autoconfigure/HotKeyBroadcastAutoConfiguration.java
@@ -1,10 +1,18 @@
-package io.github.hyshmily.hotkey.broadcast;
+package io.github.hyshmily.hotkey.autoconfigure;
 
 import com.github.benmanes.caffeine.cache.Cache;
+import io.github.hyshmily.hotkey.broadcast.BroadcastListener;
+import io.github.hyshmily.hotkey.broadcast.BroadcastProperties;
+import io.github.hyshmily.hotkey.broadcast.BroadcastPublisher;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.function.Function;
-import org.springframework.amqp.core.*;
+import org.springframework.amqp.core.AcknowledgeMode;
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.FanoutExchange;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.QueueBuilder;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;